requirements.txt
pandas
numpy
matplotlib
PyQt5
pygame
pillow